Fixed a memory leak in deserialization.
The "leak" was caused the Dictionary used for string caching (to reuse the same string reference if repeated instead of creating a new one). When disposing the ProtoReader, this cache is cleared. However clearing does not dispose all the internal data in the Dictionary, keeping the internal hastable with the size it had. While this is good if the dictionary is to be reused (by, say, deserializing another object of similar characteristics), in our application with several million strings this led to about 80MB being wasted.

The fix simply sets the dictionary to null so it can be fully collected and will be auto-recreated for subsequent deserializations.